requirements.txt
#numpy==2.3.4
#torch==2.5.0
#transformers==4.56.1
#sentence-transformers==5.1.2
fastapi==0.121.0
uvicorn[standard]==0.38.0
requests==2.32.5
pymongo==4.15.3
python-dotenv==1.0.1
scikit-learn==1.7.2
gunicorn==23.0.0



numpy==1.26.4  # Baja si 2.3.4 causa issues con Torch
torch==2.3.1+cpu  # CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu
transformers==4.41.2  # Baja si conflicts con sentence-transformers
sentence-transformers==3.0.1  # Estable, liviana
